Move FindRepositoryRootDirectory to files package
